refactor: GoodItem 컴포넌트를 src/shared/components 하위로 이동
