+++ src/shared/components/GoodItem/index.tsx
