+++ src/shared/components/GoodItem/style.css.ts
